# docs/openapi.yaml
openapi: 3.0.0
info:
  title: InnoAI API
  version: 3.1.0
  description: |
    AI Models for Every Filipino
    
    The InnoAI platform provides access to cutting-edge AI models including chat, image generation, 
    audio processing, 3D generation, video generation, and embeddings. Built for the Philippine market 
    with local currency pricing (PHP) and payment integration.
    
    ## Authentication
    
    The API supports three authentication methods:
    
    1. **JWT Bearer Token** - For customer portal authentication
    2. **API Key** - For inference endpoints (X-API-Key header)
    3. **Admin Key** - For administrative endpoints (x-admin-key header)
    
    ## Access Control Levels
    
    | Tier | Price | Requests/Day | Tokens/Month | Concurrent | Models |
    |------|-------|--------------|--------------|------------|--------|
    | Free | ₱0 | 100 | 10K | 1 | Basic only |
    | Starter | ₱499 | 1,000 | 100K | 3 | All chat |
    | Professional | ₱1,499 | 10,000 | 1M | 10 | All models |
    | Enterprise | ₱4,999 | 100,000 | 10M | 50 | All + Priority |
    
    ## Pricing (PHP)
    
    - **Chat (Boyong Opus)**: ₱465/1M input, ₱2,325/1M output
    - **Images (FLUX Pro)**: ₱3.75/image
    - **Video (Wan 480p)**: ₱4.68/second
    - **3D Generation**: Starting at ₱0.50/model
    
    ## Base URL
    
    All API requests should be made to: `https://ai-gateway.innoserver.cloud`
    
  contact:
    name: InnovateHub Support
    email: support@innovatehub.ph
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - AdminKeyAuth: []
...
paths:
...
  /api/v1/customers/api-keys:
    get:
      tags:
        - Customers
      summary: List API keys
      description: Get all API keys for the authenticated customer
      operationId: listApiKeys
      security:
        - BearerAuth: []
      responses:
        '200':
          description: API keys retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  apiKeys:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
              example:
                success: true
                apiKeys:
                  - id: key_abc123
                    name: Production Key
                    createdAt: "2026-02-21T10:30:00.000Z"
                    lastUsed: "2026-02-21T15:45:00.000Z"
                    usageCount: 1523
                    status: active
                    expiresAt: null
                  - id: key_def456
                    name: Development Key
                    createdAt: "2026-02-20T08:00:00.000Z"
                    lastUsed: null
                    usageCount: 0
                    status: active
                    expiresAt: null
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Customers
      summary: Create API key
      description: Create a new API key for making inference requests
      operationId: createApiKey
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name for the API key
                  example: Production API Key
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  apiKey:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      key:
                        type: string
                        description: The API key - copy this now, it won't be shown again
                      createdAt:
                        type: string
                        format: date-time
                      status:
                        type: string
                        enum: [active, revoked]
                  message:
                    type: string
              example:
                success: true
                apiKey:
                  id: key_xyz789
                  name: Production API Key
                  key: ik_aBcDeFgHiJkLmNoPqRsTuVwXyZ1234567890AbCdEf
                  createdAt: "2026-02-21T10:30:00.000Z"
                  status: active
                message: API key created. Copy it now - it won't be shown again!
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customers/api-keys/{keyId}:
    delete:
      tags:
        - Customers
      summary: Revoke API key
      description: Revoke an API key by its ID
      operationId: revokeApiKey
      security:
        - BearerAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
          description: The API key ID
          example: key_xyz789
      responses:
        '200':
          description: API key revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
              example:
                success: true
                message: API key revoked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customers/usage:
    get:
      tags:
        - Customers
      summary: Get usage statistics
      description: Get usage statistics for the authenticated customer
      operationId: getUsage
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            default: 30d
          description: Time period for usage data (e.g., 7d, 30d, 90d)
      responses:
        '200':
          description: Usage statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  usage:
                    type: object
                    properties:
                      totalRequests:
                        type: integer
                      totalTokens:
                        type: integer
                      totalCost:
                        type: number
                      requestsByModel:
                        type: object
                        additionalProperties:
                          type: object
                          properties:
                            requests:
                              type: integer
                            tokens:
                              type: integer
                            cost:
                              type: number
                      recentDaily:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                            requests:
                              type: integer
                            tokens:
                              type: integer
                            cost:
                              type: number
              example:
                success: true
                usage:
                  totalRequests: 5420
                  totalTokens: 2450000
                  totalCost: 1425.50
                  requestsByModel:
                    inno-ai-boyong-4.5:
                      requests: 3200
                      tokens: 1800000
                      cost: 950.00
                    flux-1.1-pro:
                      requests: 150
                      tokens: 0
                      cost: 562.50
                  recentDaily:
                    - date: "2026-02-21"
                      requests: 180
                      tokens: 85000
                      cost: 52.30
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customers/billing:
    get:
      tags:
        - Customers
      summary: Get billing information
      description: Get current billing information for the authenticated customer
      operationId: getBilling
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Billing information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  billing:
                    type: object
                    properties:
                      tier:
                        type: string
                        enum: [free, starter, professional, enterprise]
                      currentUsage:
                        type: number
                      currency:
                        type: string
                        example: PHP
                      nextBillingDate:
                        type: string
                        format: date-time
              example:
                success: true
                billing:
                  tier: professional
                  currentUsage: 1425.50
                  currency: PHP
                  nextBillingDate: "2026-03-21T00:00:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customers/billing/subscribe:
    post:
      tags:
        - Customers
      summary: Subscribe to a tier
      description: Create a checkout session for subscription upgrade
      operationId: subscribeTier
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tier:
                  type: string
                  enum: [starter, professional, enterprise]
                  description: The subscription tier to upgrade to
            example:
              tier: professional
      responses:
        '200':
          description: Checkout session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  checkoutUrl:
                    type: string
                    format: uri
                  tier:
                    type: string
                  amount:
                    type: number
              example:
                success: true
                checkoutUrl: https://checkout.directpayph.com/session/abc123
                tier: professional
                amount: 1499
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/usage:
    get:
      tags:
        - Inference
      summary: Get API key usage
      description: Get usage statistics for the current API key
      operationId: getApiKeyUsage
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Usage retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key_id:
                    type: string
                  api_key_name:
                    type: string
                  created:
                    type: string
                  last_used:
                    type: string
                    nullable: true
                  usage:
                    type: object
                    properties:
                      requests:
                        type: integer
                      tokens:
                        type: integer
                      rate_limit:
                        type: integer
                      monthly_limit:
                        type: integer
              example:
                api_key_id: key_abc123
                api_key_name: Production Key
                created: "2026-02-21T10:30:00.000Z"
                last_used: "2026-02-21T15:45:00.000Z"
                usage:
                  requests: 1523
                  tokens: 450000
                  rate_limit: 100
                  monthly_limit: 1000000
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for customer authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for inference endpoints (format ik_xxxxxxxx...)
    AdminKeyAuth:
      type: apiKey
      in: header
      name: x-admin-key
      description: Admin key for administrative endpoints
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          example: key_abc123
        name:
          type: string
          example: Production Key
        createdAt:
          type: string
          format: date-time
        lastUsed:
          type: string
          format: date-time
          nullable: true
        usageCount:
          type: integer
        status:
          type: string
          enum: [active, revoked]
        expiresAt:
          type: string
          format: date-time
          nullable: true
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: Invalid request parameters
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: Authentication required
...